Move div_rem to universal.rs
